+++ calypso_anemometer/util.py
@@ -1,9 +1,13 @@
 import asyncio
 import dataclasses
+import enum
 import functools
 import json
 import logging
 import sys
+import typing as t
+
+import click
 
 
 def setup_logging(level=logging.INFO):
@@ -48,3 +52,19 @@ def to_json(obj, pretty=True):
     if pretty:
         kwargs["indent"] = 2
     return json.dumps(obj, cls=JsonEncoderPlus, **kwargs)
+
+
+class EnumChoice(click.Choice):
+    # https://github.com/pallets/click/pull/2210
+    def __init__(self, enum_type: t.Type[enum.Enum], case_sensitive: bool = True):
+        super().__init__(
+            choices=[element.name for element in enum_type],
+            case_sensitive=case_sensitive,
+        )
+        self.enum_type = enum_type
+
+    def convert(self, value: t.Any, param: t.Optional["Parameter"], ctx: t.Optional["Context"]) -> t.Any:
+        value = super().convert(value=value, param=param, ctx=ctx)
+        if value is None:
+            return None
+        return self.enum_type[value]